+++ stripecardscan/src/androidTest/java/com/stripe/android/stripecardscan/framework/image/BitmapExtensionsTest.kt
@@ -5,15 +5,15 @@ import android.util.Size
 import androidx.core.graphics.drawable.toBitmap
 import androidx.test.filters.SmallTest
 import androidx.test.platform.app.InstrumentationRegistry
+import com.stripe.android.camera.framework.image.crop
+import com.stripe.android.camera.framework.image.scale
 import com.stripe.android.stripecardscan.test.R
-import org.junit.Ignore
 import org.junit.Test
 import kotlin.test.assertEquals
+import kotlin.test.assertNotEquals
 import kotlin.test.assertNotNull
 import kotlin.test.assertTrue
 
-@Ignore("Failing tests")
-// TODO(awush-stripe): Fix and re-enable tests
 class BitmapExtensionsTest {
 
     private val testResources = InstrumentationRegistry.getInstrumentation().context.resources
@@ -25,8 +25,9 @@ class BitmapExtensionsTest {
         val bitmap = testResources.getDrawable(R.drawable.ocr_card_numbers, null)
             .toBitmap()
         assertNotNull(bitmap)
-        assertEquals(600, bitmap.width, "Bitmap width is not expected")
-        assertEquals(375, bitmap.height, "Bitmap height is not expected")
+        // Make sure a non-empty image is read.
+        assertNotEquals(0, bitmap.width, "Bitmap width is 0")
+        assertNotEquals(0, bitmap.height, "Bitmap height is 0")
 
         // scale the bitmap
         val scaledBitmap = bitmap.scale(0.2F)
@@ -56,8 +57,9 @@ class BitmapExtensionsTest {
         val bitmap = testResources.getDrawable(R.drawable.ocr_card_numbers, null)
             .toBitmap()
         assertNotNull(bitmap)
-        assertEquals(600, bitmap.width, "Bitmap width is not expected")
-        assertEquals(375, bitmap.height, "Bitmap height is not expected")
+        // Make sure a non-empty image is read.
+        assertNotEquals(0, bitmap.width, "Bitmap width is 0")
+        assertNotEquals(0, bitmap.height, "Bitmap height is 0")
 
         // crop the bitmap
         val croppedBitmap = bitmap.crop(